Removing RunClient() func from main and use temporary file during development
